+++ api/src/main/java/org/allaymc/api/entity/component/player/EntityPlayerAttributeComponent.java
@@ -1,33 +1,12 @@
 package org.allaymc.api.entity.component.player;
 
-import com.google.common.collect.Lists;
 import org.allaymc.api.entity.component.attribute.AttributeType;
 import org.allaymc.api.entity.component.attribute.EntityAttributeComponent;
-
-import java.util.Arrays;
 
 /**
  * @author daoge_cmd
  */
 public interface EntityPlayerAttributeComponent extends EntityAttributeComponent {
-
-    /**
-     * Return an array of basic player attributes.
-     *
-     * @return array of basic player attributes
-     */
-    static AttributeType[] basicPlayerAttributes() {
-        var list = Lists.newArrayList(
-                AttributeType.PLAYER_HUNGER,
-                AttributeType.PLAYER_SATURATION,
-                AttributeType.PLAYER_EXHAUSTION,
-                AttributeType.PLAYER_EXPERIENCE_LEVEL,
-                AttributeType.PLAYER_EXPERIENCE_PROGRESS
-        );
-        list.addAll(Arrays.asList(EntityAttributeComponent.basicEntityAttributes()));
-        return list.toArray(AttributeType[]::new);
-    }
-
     /**
      * Calculates the required experience for a given level.
      *