+++ api/src/main/java/org/allaymc/api/entity/component/player/EntityPlayerChunkLoaderComponent.java
@@ -1,0 +1,10 @@
+package org.allaymc.api.entity.component.player;
+
+import org.allaymc.api.entity.component.EntityComponent;
+import org.allaymc.api.world.chunk.ChunkLoader;
+
+/**
+ * @author daoge_cmd
+ */
+public interface EntityPlayerChunkLoaderComponent extends EntityComponent, ChunkLoader {
+}