+++ api/src/main/java/org/allaymc/api/entity/interfaces/EntityPlayer.java
@@ -32,7 +32,8 @@ public interface EntityPlayer extends
         EntityPhysicsComponent,
         EntityPlayerScoreboardViewerComponent,
         EntityPlayerDebugShapeViewerComponent,
-        EntityPlayerEntityViewerComponent {
+        EntityPlayerEntityViewerComponent,
+        EntityPlayerChunkLoaderComponent {
 
     /**
      * Returns the reachable container for the given full container type.