+++ api/src/main/java/org/allaymc/api/world/WorldViewer.java
@@ -1,0 +1,10 @@
+package org.allaymc.api.world;
+
+/**
+ * WorldViewer represents an object that can view changes made in a world, such as
+ * the block changes.
+ *
+ * @author daoge_cmd
+ */
+public interface WorldViewer {
+}