+++ server/src/main/java/org/allaymc/server/entity/type/EntityTypeInitializer.java
@@ -2,7 +2,6 @@ package org.allaymc.server.entity.type;
 
 import lombok.experimental.UtilityClass;
 import org.allaymc.api.entity.component.attribute.AttributeType;
-import org.allaymc.api.entity.component.player.EntityPlayerAttributeComponent;
 import org.allaymc.api.entity.data.EntityId;
 import org.allaymc.api.entity.type.EntityTypes;
 import org.allaymc.server.entity.component.*;
@@ -10,7 +9,7 @@ import org.allaymc.server.entity.component.player.*;
 import org.allaymc.server.entity.component.projectile.*;
 import org.allaymc.server.entity.impl.*;
 
-import static org.allaymc.api.entity.component.attribute.EntityAttributeComponent.basicEntityAttributes;
+import static org.allaymc.server.entity.component.EntityAttributeComponentImpl.basicEntityAttributes;
 
 /**
  * @author daoge_cmd
@@ -75,7 +74,7 @@ public final class EntityTypeInitializer {
                 .vanillaEntity(EntityId.PLAYER)
                 .addComponent(EntityPlayerBaseComponentImpl::new, EntityPlayerBaseComponentImpl.class)
                 .addComponent(EntityPlayerNetworkComponentImpl::new, EntityPlayerNetworkComponentImpl.class)
-                .addComponent(() -> new EntityPlayerAttributeComponentImpl(EntityPlayerAttributeComponent.basicPlayerAttributes()), EntityPlayerAttributeComponentImpl.class)
+                .addComponent(() -> new EntityPlayerAttributeComponentImpl(EntityPlayerAttributeComponentImpl.basicPlayerAttributes()), EntityPlayerAttributeComponentImpl.class)
                 .addComponent(EntityPlayerContainerHolderComponentImpl::new, EntityPlayerContainerHolderComponentImpl.class)
                 .addComponent(EntityPlayerContainerViewerComponentImpl::new, EntityPlayerContainerViewerComponentImpl.class)
                 .addComponent(EntityPlayerDamageComponentImpl::new, EntityPlayerDamageComponentImpl.class)
@@ -84,6 +83,7 @@ public final class EntityTypeInitializer {
                 .addComponent(EntityPlayerScoreboardViewerComponentImpl::new, EntityPlayerScoreboardViewerComponentImpl.class)
                 .addComponent(EntityPlayerDebugShapeViewerComponentImpl::new, EntityPlayerDebugShapeViewerComponentImpl.class)
                 .addComponent(EntityPlayerEntityViewerComponentImpl::new, EntityPlayerEntityViewerComponentImpl.class)
+                .addComponent(EntityPlayerChunkLoaderComponentImpl::new, EntityPlayerChunkLoaderComponentImpl.class)
                 .build();
     }